+++ packages/client/src/components/git/index.ts
@@ -3,6 +3,7 @@ export * from "./git-branch-control";
 export * from "./git-commit-row";
 export * from "./git-create-branch-dialog";
 export * from "./git-diff.utility";
+export * from "./git-diff-control";
 export * from "./git-diff-viewer";
 export * from "./git-history";
 export * from "./git-restore-dialog";
